Creating a helper class.
